fix: resolving sonar issue
