+++ src/main/java/com/dinosaur/dinosaurexploder/controller/core/collisions/ally/AllyOrangeDinoCollision.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-package com.dinosaur.dinosaurexploder.controller.core.collisions.Player;
+package com.dinosaur.dinosaurexploder.controller.core.collisions.ally;
 
 import static com.almasb.fxgl.dsl.FXGL.onCollisionBegin;
 
@@ -12,26 +12,24 @@ import com.dinosaur.dinosaurexploder.constants.GameConstants;
 import com.dinosaur.dinosaurexploder.controller.core.GameActions;
 import com.dinosaur.dinosaurexploder.controller.core.collisions.CollisionHandlerInterface;
 import com.dinosaur.dinosaurexploder.utils.AudioManager;
-import java.util.logging.Logger;
 
-public class PlayerOrangeDinoCollision implements CollisionHandlerInterface {
+public class AllyOrangeDinoCollision implements CollisionHandlerInterface {
 
   private final GameActions gameActions;
-  private Logger logger = Logger.getLogger(getClass().getName());
 
-  public PlayerOrangeDinoCollision(GameActions gameActions) {
+  public AllyOrangeDinoCollision(GameActions gameActions) {
     this.gameActions = gameActions;
   }
 
   @Override
   public void register() {
     onCollisionBegin(
-        EntityType.PLAYER,
+        EntityType.ALLY,
         EntityType.ORANGE_DINO,
-        (player, orangeDino) -> {
+        (ally, orangeDino) -> {
+          // Maybe we could add a different for when the ally is damage
           AudioManager.getInstance().playSound(GameConstants.PLAYER_HIT_SOUND);
-          logger.info("You touched a orange dino !");
-          gameActions.damagePlayer();
+          gameActions.damageAlly();
         });
   }
 }
