+++ src/main/java/com/dinosaur/dinosaurexploder/controller/core/collisions/ally/AllyRedDinoCollision.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-package com.dinosaur.dinosaurexploder.controller.core.collisions.Player;
+package com.dinosaur.dinosaurexploder.controller.core.collisions.ally;
 
 import static com.almasb.fxgl.dsl.FXGL.onCollisionBegin;
 
@@ -12,27 +12,24 @@ import com.dinosaur.dinosaurexploder.constants.GameConstants;
 import com.dinosaur.dinosaurexploder.controller.core.GameActions;
 import com.dinosaur.dinosaurexploder.controller.core.collisions.CollisionHandlerInterface;
 import com.dinosaur.dinosaurexploder.utils.AudioManager;
-import java.util.logging.Logger;
 
-public class EnemyProjectilePlayerCollision implements CollisionHandlerInterface {
+public class AllyRedDinoCollision implements CollisionHandlerInterface {
 
   private final GameActions gameActions;
-  private Logger logger = Logger.getLogger(getClass().getName());
 
-  public EnemyProjectilePlayerCollision(GameActions gameActions) {
+  public AllyRedDinoCollision(GameActions gameActions) {
     this.gameActions = gameActions;
   }
 
   @Override
   public void register() {
     onCollisionBegin(
-        EntityType.ENEMY_PROJECTILE,
-        EntityType.PLAYER,
-        (projectile, player) -> {
+        EntityType.ALLY,
+        EntityType.RED_DINO,
+        (ally, redDino) -> {
+          // Maybe we could add a different for when the ally is damage
           AudioManager.getInstance().playSound(GameConstants.PLAYER_HIT_SOUND);
-          projectile.removeFromWorld();
-          logger.info("You got hit !\n");
-          gameActions.damagePlayer();
+          gameActions.damageAlly();
         });
   }
 }
