+++ src/main/java/com/dinosaur/dinosaurexploder/controller/core/collisions/player/EnemyProjectilePlayerCollision.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-package com.dinosaur.dinosaurexploder.controller.core.collisions.Ally;
+package com.dinosaur.dinosaurexploder.controller.core.collisions.player;
 
 import static com.almasb.fxgl.dsl.FXGL.onCollisionBegin;
 
@@ -12,24 +12,27 @@ import com.dinosaur.dinosaurexploder.constants.GameConstants;
 import com.dinosaur.dinosaurexploder.controller.core.GameActions;
 import com.dinosaur.dinosaurexploder.controller.core.collisions.CollisionHandlerInterface;
 import com.dinosaur.dinosaurexploder.utils.AudioManager;
+import java.util.logging.Logger;
 
-public class AllyRedDinoCollision implements CollisionHandlerInterface {
+public class EnemyProjectilePlayerCollision implements CollisionHandlerInterface {
 
   private final GameActions gameActions;
+  private Logger logger = Logger.getLogger(getClass().getName());
 
-  public AllyRedDinoCollision(GameActions gameActions) {
+  public EnemyProjectilePlayerCollision(GameActions gameActions) {
     this.gameActions = gameActions;
   }
 
   @Override
   public void register() {
     onCollisionBegin(
-        EntityType.ALLY,
-        EntityType.RED_DINO,
-        (ally, redDino) -> {
-          // Maybe we could add a different for when the ally is damage
+        EntityType.ENEMY_PROJECTILE,
+        EntityType.PLAYER,
+        (projectile, player) -> {
           AudioManager.getInstance().playSound(GameConstants.PLAYER_HIT_SOUND);
-          gameActions.damageAlly();
+          projectile.removeFromWorld();
+          logger.info("You got hit !\n");
+          gameActions.damagePlayer();
         });
   }
 }
